Replace watchOffset option with individual top/bottom/left/right options
Having only one offset value can cause issues in more advanced use
cases, especially if watchOnce is false and your watched element is
small. For example, watching a 1x1 element with an offset of 500 at
certain viewports would report that it was not in view due to the
offset being accounted for in the x axis. Having individualized options
is much more powerful and provides more expected behavior.